create state pkg
