+++ state/state.go
@@ -1,0 +1,24 @@
+package state
+
+import (
+	"context"
+
+	"github.com/ava-labs/avalanchego/x/merkledb"
+)
+
+type Immutable interface {
+	GetValue(ctx context.Context, key []byte) (value []byte, err error)
+}
+
+type Mutable interface {
+	Immutable
+
+	Insert(ctx context.Context, key []byte, value []byte) error
+	Remove(ctx context.Context, key []byte) error
+}
+
+type View interface {
+	Immutable
+
+	NewViewFromMap(ctx context.Context, changes map[string]*merkledb.ChangeOp, copyBytes bool) (merkledb.TrieView, error)
+}
